Allow deserialization of types originally only referenced via interface
