+++ src/SampleServerClasses/IMyDto.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleServerClasses
+{
+	public interface IMyDto
+	{
+		string Name { get; }
+		int Id { get; }
+
+		IMyDto Next { get; set; }
+	}
+}
